better cache

+++ rplugin/python3/fancy_completion/server/types.py
@@ -19,7 +19,7 @@ class FuzzyOptions:
 class CacheOptions:
     short_name: str
     band_size: int
-    limit: int
+    limit: float
     source_name: str = "cache"
